ci/requirements.txt
anyio==3.6.2
attrs==23.1.0
backoff==2.2.1
beartype==0.14.0
cattrs==22.2.0
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
dagger-io==0.5.4
gql==3.4.1
graphql-core==3.2.3
h11==0.14.0
httpcore==0.17.1
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
packaging==23.1
platformdirs==3.5.1
pluggy==1.0.0
Pygments==2.15.1
pytest==7.3.1
rich==13.3.5
shellingham==1.5.0.post1
sniffio==1.3.0
typer==0.9.0
typing_extensions==4.5.0
yarl==1.9.2
